requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
pymupdf
chromadb
openai
tiktoken
requests
arxiv
langchain-text-splitters